Allow scrolling
This was a confusing one.  We have a non-logged-in route which doesn't
render what I had expected.  The @scroll=true ensures we can scroll on
the page, which means Metis can show a page with more than 4 properties.
A bit of a CSS design challenge but looks like it was supported.